Updated to support SQLite. Please check the Project progress.txt file for details

+++ AgeRanger.Data/Contracts/IBetweenMinMaxRule.cs
@@ -2,7 +2,7 @@
 {
   public interface IBetweenMinMaxRule<T> : IRule<T> where T : IUser
   {
-    int Min { get; set; }
-    int Max { get; set; }
+    long Min { get; set; }
+    long Max { get; set; }
   }
 }
