+++ AgeRanger.Data/Contracts/IRepository.cs
@@ -5,10 +5,10 @@
   public interface IRepository<T> where T : class
   {
     IQueryable<T> GetAll();
-    T GetById(int id);
+    T GetById(long id);
     void Add(T entity);
     void Update(T entity);
     void Delete(T entity);
-    void Delete(int id);
+    void Delete(long id);
   }
 }
